Fix the example to act as a server (#100)
Fixes a bug mentioned in #50

Both the client and server examples were shown with
`WebSocketChannel.connect`. Update the server to bind the `HttpServer`
instead of trying to connect to a server that hasn't been started.

Add a `client.close()` call to the example so the VM does not continue
running.